Draft of IPython Notebook Example
This simple example demonstrates a workflow for using 1) PDBFixer to
generate a usable PDB file for MD, 2) OpenMM for running 2 MD
simulations, 3) MDTraj to compute RMSD and dihedral angles to compare
the two trajectories.  For now, PDBFixer is included in the directory to
make it easy to launch from inside the notebook.